Refactors Tfl client not to use deprecated URI.encode
Ruby > 2.7 deprecated the URI.encode method. This was showing up as a warning in a number of places. This refactors the library to instead use the URI.encode_www_form_component method. Additionally the /CycleSuperhighways endpoint has been deprecated by TfL. This removes support for this API endpoint. All libraries have been bumped to support ruby 2.7+.